Fix server-mode connection reuse error
Fix an issue that caused issues when Xinu was acting as a server. When
connections to clients expired, the refcount of the associated TCB was
not fully decremented (that is, it only decremented to 1 instead of 0 as
would be expected).

Obviously, tcplisten allocated a TCB and incremented its refcount to 1
when a SYN was received from a client. However, for some reason, the
refcount was also incremented in tcpsynrcvd after receiving the client's
ACK of our SYN/ACK, just before marking the connection as ESTABLISHED. A
comment noted that this was done "for the system," but there is no
reason that it should have been necessary.